Changes to wctl, mild cleanup
